add authentication behaviour for answer

+++ app/Http/Controllers/AnswersController.php
@@ -56,6 +56,13 @@ class AnswersController extends Controller
      */
     public function store(Request $request)
     {
+        $isMember = Auth::guard('member')->user() != null;
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
+
+         // if not member and admin then cannot edit this answer
+        if(!($isMember || $isAdmin))
+            return redirect('/');
+
         $this->validate($request, [
             'body' => 'required',
         ]);
@@ -64,6 +71,11 @@ class AnswersController extends Controller
         $answer->question_id = $request->input('question_id');
         $answer->body = $request->input('body');
         $answer->user_id = auth()->user()->id;
+        if ($isAdmin) {
+            $answer->is_admin = 1;
+        } else {
+            $answer->is_admin = 0;
+        }
         $answer->save();
 
         return redirect('/admin/questions')->with('success', 'Answer Saved');
@@ -93,9 +105,24 @@ class AnswersController extends Controller
      */
     public function edit($id)
     {
+        $isMember = Auth::guard('member')->user() != null;
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
+
+        // if not member and admin then cannot edit this question
+        if(!($isMember || $isAdmin))
+            return redirect('/');
+        
         $answer = Answer::find($id);
+        $member = Auth::guard('member')->user();
+
         if ($answer !== null) {
-            return view('admin.editanswer')->with('answer', $answer);
+            // if admin can edit all kinds of answer
+            // if not admin can only edit his own answer
+            if ($isAdmin || ($answer->user()->id == $member->id && $answer->is_admin == 0)) {
+                return view('admin.editanswer')->with('answer', $answer);
+            } else {
+                return redirect('/admin/questions')->with('error', 'You can not edit this answer.');
+            }
         } else {
             return abort(404);
         }
@@ -146,15 +173,27 @@ class AnswersController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $isMember = Auth::guard('member')->user() != null;
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
+
+        // if not member and admin then cannot edit this question
+        if(!($isMember || $isAdmin))
+            return redirect('/');
+
         $this->validate($request, [
             'body' => 'required',
         ]);
 
         $answer = Answer::find($id);
-        $answer->body = $request->input('body');
-        $answer->save();
+        $member = Auth::guard('member')->user();
 
-        return redirect('/admin/questions')->with('success', 'Answer Updated');
+        if ($isAdmin || ($answer->user()->id == $member->id && $answer->is_admin == 0)) {
+            $answer->body = $request->input('body');
+            $answer->save();
+            return redirect('/admin/questions')->with('success', 'Answer Updated');
+        } else {
+            return redirect('/admin/questions')->with('error', 'You can not edit this answer.');
+        }
     }
 
     /**
@@ -165,9 +204,27 @@ class AnswersController extends Controller
      */
     public function destroy($id)
     {
-        $answer = Answer::find($id);
-        $answer->delete();
+        $isMember = Auth::guard('member')->user() != null;
+        $isAdmin = Auth::user() != null && Auth::user()->IsAdmin == 1;
 
-        return redirect('/admin/questions')->with('error', 'Answer Deleted');
+        // if not member and admin then cannot delete this answer
+        if(!($isMember || $isAdmin))
+            return redirect('/');
+        
+        $answer = Answer::find($id);
+        $member = Auth::guard('member')->user();
+
+        if ($answer !== null) {
+            // if admin can delete all kinds of answer
+            // if not admin can only delete his own answer
+            if ($isAdmin || ($answer->user()->id == $member->id && $answer->is_admin == 0)) {
+                $answer->delete();
+                return redirect('/admin/questions')->with('error', 'Answer Deleted');
+            } else {
+                return redirect('/admin/questions')->with('error', 'You can not delete this answer.');
+            }
+        } else {
+            return abort(404);
+        }
     }
 }
